Prevent station from engaging when...
the passenger has already passed the station as determined by states in
its state history.

Also, changed the passenger constructor sig to take an options hash which
accepts a ticket and a state_history. This allows passenger construction
with a pre-existing state_history. This allows current_state
modificaiton and safe rerun through the route capability.